test(src): move inline test modules to tests dir (closes #145)
Moves all 6 inline #[cfg(test)] modules out of src/worker/ and src/daemon/ into tests/ (4 new [[test]] targets, 2 appends). Adds test seams (truncate_body_for_tests, parse_starttime_from_stat_for_tests, outcome_for_class_for_tests, map_phase_to_outcome_for_tests, extract_http_status_for_tests) and extends FakeClock cfg_attr to its impls.
