+++ src/worker/supervisor/mod.rs
@@ -71,3 +71,8 @@ pub use framing::*;
 pub use heartbeat::*;
 pub use outcome_transcript::*;
 pub use process_lifecycle::*;
+
+// Test seam: re-export the synthetic-stat parser so integration
+// tests can assert the field-22 contract without owning a runtime.
+#[doc(hidden)]
+pub use self::process_lifecycle::parse_starttime_from_stat_for_tests;
