+++ tests/daemon/orchestration_active_run_inline_test.rs
@@ -1,0 +1,371 @@
+//! Integration tests for the orchestration active-run guard
+//! (`src/daemon/orchestration/active_run.rs`). Moved out of the
+//! inline `#[cfg(test)]` module per AGENTS.md.
+
+use caduceus::infra::config::Config;
+use caduceus::orchestration::{classify_error, Clock, FailureClass, FakeClock, SystemClock};
+use caduceus::queue::{ClaimFileBody, ClaimToken, CLAIM_FILE_VERSION};
+use caduceus::{CaduceusError, IssueKey};
+use chrono::{DateTime, Utc};
+use std::path::PathBuf;
+use std::sync::Arc;
+
+fn cfg() -> Config {
+    Config::test_defaults(std::path::Path::new("/tmp"))
+}
+
+fn dummy_claim() -> ClaimToken {
+    ClaimToken::for_test(
+        std::env::temp_dir().join("caduceus-orchestration-tests"),
+        "deadbeef00",
+        "RUNID",
+    )
+}
+
+#[test]
+fn classify_error_maps_variants() {
+    // Cancellation
+    let err = CaduceusError::Cancelled;
+    assert_eq!(classify_error(&err), FailureClass::Cancellation);
+
+    // Rate limit
+    let err = CaduceusError::RateLimited {
+        reset_at: 12345,
+        remaining: 0,
+        limit: Some(5000),
+    };
+    assert_eq!(
+        classify_error(&err),
+        FailureClass::RateLimit { reset_at: 12345 }
+    );
+
+    // Worker-attributable
+    let err = CaduceusError::Worker {
+        context: "result",
+        stderr: "schema mismatch".to_string(),
+    };
+    assert_eq!(classify_error(&err), FailureClass::Worker);
+
+    let err = CaduceusError::Other("voice: forbidden term".to_string());
+    assert_eq!(classify_error(&err), FailureClass::Worker);
+
+    // Infrastructure
+    let err = CaduceusError::Config("bad worker command".to_string());
+    assert_eq!(classify_error(&err), FailureClass::Infrastructure);
+
+    let err = CaduceusError::TokenResolution("gh not found".to_string());
+    assert_eq!(classify_error(&err), FailureClass::Infrastructure);
+
+    // HTTP transport — infrastructure
+    let io_err = std::io::Error::new(std::io::ErrorKind::ConnectionReset, "reset");
+    let err: CaduceusError = io_err.into();
+    assert_eq!(classify_error(&err), FailureClass::Infrastructure);
+
+    // Pool saturated — infrastructure
+    let err = CaduceusError::PoolSaturated {
+        current_depth: 5,
+        max_depth: 10,
+    };
+    assert_eq!(classify_error(&err), FailureClass::Infrastructure);
+
+    // Repository exclusion held — infrastructure
+    let err = CaduceusError::RepositoryExclusionHeld {
+        repo_key: "owner/repo".into(),
+    };
+    assert_eq!(classify_error(&err), FailureClass::Infrastructure);
+
+    // Drain timeout — cancellation
+    let err = CaduceusError::DrainTimeout {
+        timed_out_run_ids: vec!["run-1".into()],
+    };
+    assert_eq!(classify_error(&err), FailureClass::Cancellation);
+
+    // SymlinkedStorageRoot — infrastructure
+    let err = CaduceusError::SymlinkedStorageRoot {
+        path: PathBuf::from("/tmp/link"),
+    };
+    assert_eq!(classify_error(&err), FailureClass::Infrastructure);
+
+    // WorktreeReuseAfterFailure — worker
+    let err = CaduceusError::WorktreeReuseAfterFailure {
+        run_id: "deadbeef".into(),
+        worktree_path: PathBuf::from("/tmp/failed"),
+        last_state: "Failed".into(),
+    };
+    assert_eq!(classify_error(&err), FailureClass::Worker);
+
+    // ModeNotPreserved — infrastructure
+    let err = CaduceusError::ModeNotPreserved {
+        path: PathBuf::from("/tmp/x"),
+        expected: 0o700,
+        observed: 0o755,
+    };
+    assert_eq!(classify_error(&err), FailureClass::Infrastructure);
+}
+
+#[test]
+fn classify_error_is_exhaustive_at_compile_time() {
+    // Every CaduceusError variant is classified. If a new
+    // variant is added without a `classify_error` arm,
+    // this match will fail to compile.
+    let variants: [CaduceusError; 40] = [
+        CaduceusError::Config("x".into()),
+        CaduceusError::Io(std::io::Error::other("x")),
+        CaduceusError::Json(serde_json::from_str::<u8>("not-a-number").unwrap_err()),
+        CaduceusError::Yaml(serde_yaml::from_str::<u8>(": x").unwrap_err()),
+        CaduceusError::Worker {
+            context: "spawn",
+            stderr: "x".into(),
+        },
+        CaduceusError::Worktree {
+            context: "create",
+            stderr: "x".into(),
+        },
+        CaduceusError::Queue {
+            context: "claim",
+            stderr: "x".into(),
+        },
+        CaduceusError::Push {
+            context: "push",
+            stderr: "x".into(),
+        },
+        CaduceusError::PushCollision {
+            branch: "b".into(),
+            remote_oid: "r".into(),
+            local_oid: "l".into(),
+        },
+        CaduceusError::StateCorrupt {
+            path: PathBuf::from("/tmp/x"),
+            message: "x".into(),
+        },
+        CaduceusError::Git {
+            operation: "commit",
+            stderr: "x".into(),
+        },
+        CaduceusError::GitHubApi {
+            status: 500,
+            message: "x".into(),
+        },
+        CaduceusError::RateLimited {
+            reset_at: 1,
+            remaining: 0,
+            limit: None,
+        },
+        CaduceusError::TokenResolution("x".into()),
+        CaduceusError::Cancelled,
+        CaduceusError::Other("x".into()),
+        // Http is exercised via the Io case above; we
+        // synthesise an Http variant by running reqwest's
+        // error path indirectly. For the compile-time
+        // guard the variant list is the actual concern.
+        CaduceusError::Worker {
+            context: "http",
+            stderr: "transport".into(),
+        },
+        CaduceusError::LeadershipContended {
+            context: "acquire",
+            stderr: "contended".into(),
+        },
+        CaduceusError::LeaseStale {
+            context: "renew",
+            stderr: "expired".into(),
+        },
+        CaduceusError::FencingTokenRegression {
+            issue_key: "owner/repo#1".into(),
+            stale_token: 1,
+            current_token: 3,
+        },
+        CaduceusError::PoolSaturated {
+            current_depth: 1,
+            max_depth: 2,
+        },
+        CaduceusError::RepositoryExclusionHeld {
+            repo_key: "owner/repo".into(),
+        },
+        CaduceusError::DrainTimeout {
+            timed_out_run_ids: vec!["run-1".into()],
+        },
+        CaduceusError::CircuitOpen {
+            scope: "provider",
+            scope_id: "github".into(),
+            retry_after: 1800,
+            probe_in_flight: false,
+        },
+        CaduceusError::MaxDegradedAgeExceeded {
+            scope: "repository",
+            scope_id: "owner/repo".into(),
+            opened_at: 1000000,
+        },
+        CaduceusError::SymlinkedStorageRoot {
+            path: PathBuf::from("/tmp/link"),
+        },
+        CaduceusError::WorktreeReuseAfterFailure {
+            run_id: "deadbeef".into(),
+            worktree_path: PathBuf::from("/tmp/failed"),
+            last_state: "Failed".into(),
+        },
+        CaduceusError::ModeNotPreserved {
+            path: PathBuf::from("/tmp/x"),
+            expected: 0o700,
+            observed: 0o755,
+        },
+        CaduceusError::OciCliNotFound {
+            cli: "docker".into(),
+        },
+        CaduceusError::OciEngineUnavailable {
+            detail: "Cannot connect".into(),
+        },
+        CaduceusError::OciMismatchedCliVersion {
+            detail: "too old".into(),
+        },
+        CaduceusError::OciPullFailed {
+            image: "img".into(),
+            stderr: "pull failed".into(),
+        },
+        CaduceusError::OciCreateFailed {
+            context: "create",
+            stderr: "err".into(),
+        },
+        CaduceusError::OciStartFailed {
+            context: "start",
+            stderr: "err".into(),
+        },
+        CaduceusError::OciWaitFailed {
+            context: "wait",
+            stderr: "err".into(),
+        },
+        CaduceusError::OciStopFailed {
+            context: "stop",
+            stderr: "err".into(),
+        },
+        CaduceusError::OciRemoveFailed {
+            context: "remove",
+            stderr: "err".into(),
+        },
+        CaduceusError::OciUndeclaredMount {
+            path: "/tmp/x".into(),
+        },
+        CaduceusError::OciSecretLeakSuspected {
+            path: "/tmp/x".into(),
+        },
+        CaduceusError::OciSecretLeakDetected {
+            run_id: "run-1".into(),
+        },
+    ];
+    for v in &variants {
+        let _class = classify_error(v);
+    }
+    // Http is also covered by the match arms even though
+    // we don't synthesise one here.
+}
+
+#[test]
+fn failure_class_predicates() {
+    let worker = FailureClass::Worker;
+    assert!(worker.counts_against_retry_budget());
+    assert!(!worker.must_persist_rate_limit());
+    assert!(!worker.is_cancellation());
+
+    let infra = FailureClass::Infrastructure;
+    assert!(!infra.counts_against_retry_budget());
+    assert!(!infra.must_persist_rate_limit());
+    assert!(!infra.is_cancellation());
+
+    let rate = FailureClass::RateLimit { reset_at: 100 };
+    assert!(!rate.counts_against_retry_budget());
+    assert!(rate.must_persist_rate_limit());
+    assert!(!rate.is_cancellation());
+
+    let cancel = FailureClass::Cancellation;
+    assert!(!cancel.counts_against_retry_budget());
+    assert!(!cancel.must_persist_rate_limit());
+    assert!(cancel.is_cancellation());
+}
+
+#[test]
+fn claim_token_key_accessor_returns_placeholder() {
+    // The convenience accessor compiles and returns a
+    // reference; the orchestrator's higher-level code
+    // keeps its own typed `IssueKey` alongside the guard.
+    let claim = dummy_claim();
+    let key = claim.key();
+    assert_eq!(key.number, 0);
+    assert!(key.owner.is_empty());
+}
+
+#[test]
+fn claim_file_body_round_trip() {
+    // Sanity check that the queue module's ClaimFileBody
+    // round-trips so the orchestrator can rehydrate a
+    // claim token if needed.
+    let key = IssueKey::parse("owner/repo#1").expect("key");
+    let started_at: DateTime<Utc> = "2026-01-01T00:00:00Z".parse().expect("timestamp");
+    let body = ClaimFileBody {
+        version: CLAIM_FILE_VERSION,
+        key: key.clone(),
+        run_id: "RUNID".to_string(),
+        pid: 42,
+        process_start_identity: "boot-1/100".to_string(),
+        started_at,
+        worktree_path: None,
+    };
+    let serialized = serde_json::to_string(&body).unwrap();
+    let parsed: ClaimFileBody = serde_json::from_str(&serialized).unwrap();
+    assert_eq!(parsed.version, CLAIM_FILE_VERSION);
+    assert_eq!(parsed.key, key);
+    assert_eq!(parsed.run_id, "RUNID");
+}
+
+#[test]
+fn system_clock_returns_recent_utc() {
+    let before = Utc::now();
+    let now = SystemClock.now();
+    let after = Utc::now();
+    assert!(now >= before);
+    assert!(now <= after);
+}
+
+#[test]
+fn services_production_helper_compiles() {
+    // The constructor wires the four adapters. We can't
+    // call services.executor.run here because that
+    // would spawn a worker; the test only verifies the
+    // type compiles and the field accessors work.
+    let clock: Arc<dyn Clock> = Arc::new(SystemClock);
+    let _ = clock.now();
+    let _cfg = cfg();
+}
+
+#[test]
+fn fake_clock_default_starts_at_epoch() {
+    let fc = FakeClock::new();
+    assert_eq!(fc.now_unix(), 0);
+}
+
+#[test]
+fn fake_clock_advance_works() {
+    let fc = FakeClock::new();
+    fc.advance(100);
+    assert_eq!(fc.now_unix(), 100);
+}
+
+#[test]
+fn fake_clock_set_works() {
+    let fc = FakeClock::new();
+    fc.set(999);
+    assert_eq!(fc.now_unix(), 999);
+}
+
+#[test]
+fn fake_clock_clones_share_time() {
+    let fc = FakeClock::new();
+    let fc2 = fc.clone();
+    fc.advance(50);
+    assert_eq!(fc2.now_unix(), 50);
+}
+
+#[test]
+fn fake_clock_now_returns_correct_datetime() {
+    let fc = FakeClock::at(946684800); // 2000-01-01T00:00:00Z
+    assert_eq!(fc.now().timestamp(), 946684800);
+}
